+++ testing_migrations/phoenix/20200321142525_rename_column.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Phoenix\TestingMigrations;
+
+use Phoenix\Migration\AbstractMigration;
+
+class RenameColumn extends AbstractMigration
+{
+    public function up(): void
+    {
+        $this->table('table_4')->renameColumn('identifier', 'id')->save();
+    }
+
+    protected function down(): void
+    {
+        $this->table('table_4')->renameColumn('id', 'identifier')->save();
+    }
+}